Renamed column

+++ Controllers/MovieController.cs
@@ -7,8 +7,10 @@ namespace MovieMania.Controllers
     public class MovieController : Controller
     {
         private readonly ApplicationDbContext _db;
+        [Obsolete]
         private readonly IHostingEnvironment _hostingEnvironment;
 
+        [Obsolete]
         public MovieController(ApplicationDbContext db, IHostingEnvironment hostingEnvironment)
         {
             _db = db;
@@ -32,7 +34,7 @@ namespace MovieMania.Controllers
                 MovieModel newMovie = new()
                 {
                     Name = model.Name,
-                    PhotoPath = uniqueFileName ?? "avatar/noimage.jpg",
+                    PhotoName = uniqueFileName,
                     Author = model.Author,
                     Description = model.Description,
                     Duration = model.Duration,
@@ -41,6 +43,10 @@ namespace MovieMania.Controllers
 
                 _db.Add(newMovie);
                 _db.SaveChanges();
+            }
+            else
+            {
+                return View();
             }
 
             return RedirectToAction(nameof(HomeController.Index), "Home");
@@ -53,7 +59,7 @@ namespace MovieMania.Controllers
 
             if (movie == null)
             {
-                return NotFound();
+                return View("NotFound");
             }
 
             MovieEditViewModel editMovie = new()
@@ -64,7 +70,7 @@ namespace MovieMania.Controllers
                 Description = movie.Description,
                 Duration = movie.Duration,
                 ReleaseDate = movie.ReleaseDate,
-                ExistingPhotoPath = movie.PhotoPath
+                ExistingPhotoPath = movie.PhotoName
             };
 
             return View(editMovie);
@@ -84,9 +90,9 @@ namespace MovieMania.Controllers
                     movie.Description = model.Description;
                     movie.Duration = model.Duration;
                     movie.ReleaseDate = model.ReleaseDate;
-                    movie.PhotoPath = ProcessUploadFile(model);
+                    movie.PhotoName = ProcessUploadFile(model);
 
-                    if (model.Photo != null)
+                    if (model.Photo != null && !string.IsNullOrWhiteSpace(model.ExistingPhotoPath))
                     {
                         string filePath = Path.Combine(_hostingEnvironment.WebRootPath, "image", model.ExistingPhotoPath);
                         if (System.IO.File.Exists(filePath))
@@ -105,13 +111,14 @@ namespace MovieMania.Controllers
 
         private string ProcessUploadFile(MovieCreateViewModel model)
         {
-            string uniqueFileName = null;
+            string uniqueFileName = string.Empty;
             if (model.Photo != null)
             {
                 string uploadsFolder = Path.Combine(_hostingEnvironment.WebRootPath, "image");
                 uniqueFileName = Guid.NewGuid().ToString() + "_" + model.Photo.FileName;
                 string filePath = Path.Combine(uploadsFolder, uniqueFileName);
-                model.Photo.CopyTo(new FileStream(filePath, FileMode.Create));
+                using var fileStream = new FileStream(filePath, FileMode.Create);
+                model.Photo.CopyTo(fileStream);
             }
 
             return uniqueFileName;
@@ -123,13 +130,17 @@ namespace MovieMania.Controllers
             var movie = _db.Movies.FirstOrDefault(x => x.Id == id);
             if (movie == null)
             {
-                return NotFound();
+                return View("NotFound");
             }
-            var filePath = Path.Combine(_hostingEnvironment.WebRootPath, "image", movie.PhotoPath);
-            if (System.IO.File.Exists(filePath))
+            if (!String.IsNullOrWhiteSpace(movie.PhotoName))
             {
-                System.IO.File.Delete(filePath);
+                var filePath = Path.Combine(_hostingEnvironment.WebRootPath, "image", movie.PhotoName);
+                if (System.IO.File.Exists(filePath))
+                {
+                    System.IO.File.Delete(filePath);
+                }
             }
+
             _db.Remove(movie);
             _db.SaveChanges();
